rename packages / classes / functions + mark things as internal
